+++ NEWS.md
@@ -1,9 +1,10 @@
 # Version 0.0 - Release Notes
 
 * [0.2.4] 
-    * released 2023-10-12 
+    * released 2023-11-14 
     * added functions: pls_fit_species 
-    * added draft rsp_pls object class 
+    * added draft rsp_pls and rsp_x object classes 
+    * added draft sp_build_rsp_x  
     * moved from importing data.table to data.table::as.data.table, etc... 
 
 * [0.2.3]